refactor: code logic optimization
