--- README.md
@@ -1,2 +1,0 @@
-# [forecast-weather](https://vermillion-souffle-27bbf8.netlify.app/)
-new project
